refactor: refactor components, add feed back
