+++ package.json
@@ -8,7 +8,6 @@
     "enzyme": "^3.8.0",
     "enzyme-adapter-react-16": "^1.7.1",
     "history": "^4.7.2",
-    "http-proxy-middleware": "^0.19.1",
     "prop-types": "^15.6.2",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
