+++ src/reducers/index.js
@@ -1,9 +1,9 @@
 import { combineReducers } from "redux";
 import { connectRouter } from "connected-react-router";
 
-import registration from "./registrationReducer";
+import registrationReducer from "./registrationReducer";
 
 export default history => combineReducers({
   router: connectRouter(history),
-  registration
+  registrationReducer
 });
